requirements.txt
dify-plugin>=0.3.0
pillow>=10.0.0
rapidocr-onnxruntime>=1.3.0
numpy>=1.24.0
python-dotenv>=0.19.0